Move ToDigestInput impl to their respective modules
